Improved Registration, functional game launch, game crashes when packet is sent
