+++ src/main/java/com/freddy/craftyconcoctions/CraftyConcoctionsClient.java
@@ -2,7 +2,7 @@ package com.freddy.craftyconcoctions;
 
 import com.freddy.craftyconcoctions.block.ModBlockEntities;
 import com.freddy.craftyconcoctions.block.witch_cauldron.WitchCauldronBlockEntityRenderer;
-import com.freddy.craftyconcoctions.networking.ModMessagesClient;
+import com.freddy.craftyconcoctions.networking.ModMessages;
 import net.fabricmc.api.ClientModInitializer;
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactories;
 
@@ -11,8 +11,8 @@ public class CraftyConcoctionsClient implements ClientModInitializer
     @Override
     public void onInitializeClient()
     {
-        ModMessagesClient.registerS2Creceive();
-        ModMessagesClient.registerC2Ssend();
+        ModMessages.registerClientPayloadTypes();
+        ModMessages.registerS2Creceive();
 
         BlockEntityRendererFactories.register(ModBlockEntities.WITCH_CAULDRON_BE, WitchCauldronBlockEntityRenderer::new);
     }
