+++ src/main/java/com/freddy/craftyconcoctions/networking/payload/C2SPotionNameEditPayload.java
@@ -1,6 +1,6 @@
 package com.freddy.craftyconcoctions.networking.payload;
 
-import com.freddy.craftyconcoctions.networking.ModMessagesClient;
+import com.freddy.craftyconcoctions.networking.ModMessages;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.network.RegistryByteBuf;
 import net.minecraft.network.codec.PacketCodec;
@@ -9,7 +9,7 @@ import net.minecraft.network.packet.CustomPayload;
 
 public record C2SPotionNameEditPayload(String name) implements CustomPayload
 {
-    public static final CustomPayload.Id<C2SPotionNameEditPayload> ID = new CustomPayload.Id<>(ModMessagesClient.C2S_POTION_NAME_EDIT_ID);
+    public static final CustomPayload.Id<C2SPotionNameEditPayload> ID = new CustomPayload.Id<>(ModMessages.C2S_POTION_NAME_EDIT_ID);
 
     public static final PacketCodec<RegistryByteBuf, C2SPotionNameEditPayload> CODEC = PacketCodec.tuple(
             PacketCodecs.STRING, C2SPotionNameEditPayload::name,
